pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "braid-dspy"
version = "0.1.0"
description = "BRAID (Bounded Reasoning for Autonomous Inference and Decisions) integration for DSPy framework"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Braid-DSPy Contributors"}
]
keywords = ["dspy", "braid", "reasoning", "llm", "ai", "mermaid", "grd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "dspy-ai>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
    "sphinx-copybutton>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/braid-py"
Documentation = "https://github.com/yourusername/braid-py#readme"
Repository = "https://github.com/yourusername/braid-py"
Issues = "https://github.com/yourusername/braid-py/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["braid*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

